+++ tests/unit/Type/IterableTypeTest.php
@@ -59,7 +59,7 @@ final class IterableTypeTest extends TypeTest
         yield [Type\iterable(Type\array_key(), Type\int()), 'iterable<array-key, int>'];
         yield [Type\iterable(Type\array_key(), Type\string()), 'iterable<array-key, string>'];
         yield [
-            Type\iterable(Type\array_key(), Type\object(Iter\Iterator::class)),
+            Type\iterable(Type\array_key(), Type\instance_of(Iter\Iterator::class)),
             'iterable<array-key, Psl\Iter\Iterator>'
         ];
     }