+++ tests/unit/Type/NonEmptyDictTypeTest.php
@@ -96,7 +96,7 @@ final class NonEmptyDictTypeTest extends TypeTest
         yield [Type\non_empty_dict(Type\array_key(), Type\int()), 'non-empty-dict<array-key, int>'];
         yield [Type\non_empty_dict(Type\array_key(), Type\string()), 'non-empty-dict<array-key, string>'];
         yield [
-            Type\non_empty_dict(Type\array_key(), Type\object(Iter\Iterator::class)),
+            Type\non_empty_dict(Type\array_key(), Type\instance_of(Iter\Iterator::class)),
             'non-empty-dict<array-key, Psl\Iter\Iterator>'
         ];
     }